refactored some twig extensions to runtime extensions
